# webserver/angular/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST ",
        "version": "1.0.0",
        "title": "Swagger MyNerdOffice",
        "contact": {
            "email": "saverio.tosi@studenti.unicam.it"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "192.168.1.104:8000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "basicAuth": {
            "type": "basic",
            "description": "HTTP Basic Authentication. Works over `HTTP` and `HTTPS`"
        }
    },
    "paths": {
...
        "/device": {
            "get": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "device"
                ],
                "summary": "Get array of Device",
                "description": "",
                "operationId": "getDevices",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Device"
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "device"
                ],
                "summary": "Create a new device Device",
                "description": "",
                "operationId": "addDevices",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "type",
                        "description": "type of device",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "io",
                        "description": "input or output",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "permission",
                        "description": "permission",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "properties",
                        "description": "Properties",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            }
        },
        "/device/{id}": {
            "get": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "device"
                ],
                "summary": "Get one Device",
                "description": "",
                "operationId": "getDevice",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id that needs to be fetched.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Device"
                        }
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "device"
                ],
                "summary": "Edit one Device",
                "description": "",
                "operationId": "addDevice",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of Device that needs to be updated.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "automatic",
                        "description": "Set the type of operation",
                        "required": true,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Device"
                        }
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "device"
                ],
                "summary": "Delete Device",
                "description": "",
                "operationId": "deleteDevice",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of Device that needs to be deleted.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Authorization required"
                    },
                    "404": {
                        "description": "Device not found"
                    }
                }
            }
        },
...
        "/device/output": {
            "get": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "device"
                ],
                "summary": "Get array of output Devices",
                "description": "",
                "operationId": "getUnregisteredDevices",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Device"
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            }
        },
        "/beacon": {
            "get": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "beacon"
                ],
                "summary": "Get array of beacons",
                "description": "",
                "operationId": "getBeacons",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Beacon"
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "beacon"
                ],
                "summary": "Add a new beacon",
                "description": "",
                "operationId": "postBeacon",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "formData",
                        "description": "The uuid of the beacon.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "major",
                        "in": "formData",
                        "description": "The major id of the beacon.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "minor",
                        "in": "formData",
                        "description": "The minor of the beacon.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            }
        },
        "/beacon/{id}": {
            "get": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "beacon"
                ],
                "summary": "Get beacon by ID",
                "description": "",
                "operationId": "getBeacon",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id that needs to be fetched.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Beacon"
                        }
                    },
                    "401": {
                        "description": "Authorization require"
                    },
                    "404": {
                        "description": "Beacon not found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "beacon"
                ],
                "summary": "Delete a beacon",
                "description": "",
                "operationId": "deleteBeacon",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the beacon to delete.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Device": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "io": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "permission": {
                    "type": "integer"
                },
                "id_GPIO": {
                    "type": "integer"
                },
                "id_ibeacon": {
                    "type": "integer"
                },
                "properties": {
                    "type": "object"
                }
            }
        },
        "Beacon": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "minor": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                }
            }
        },
...
    }
}